Renamed dlite-print.c to dlite-json.c
