+++ src/CMakeLists.txt
@@ -42,7 +42,7 @@ set(sources
   dlite-mapping-plugins.c
   dlite-codegen.c
   dlite-getlicense.c
-  dlite-print.c
+  dlite-json.c
   getuuid.c
   pathshash.c
   triple.c
